phase-36: T7 agent b3 — func_8016CBC0 closed from a residual of 55 (128 bodies), and its move toolified as R20
- the move: narrow every local in the counter's def-use chain together — the counter, its +/-1 temp and the copy-back —
  and do it for BOTH chains at once. Four instructions were MISSING, not miscoloured, three of them the moves the $0 pin
  was faking. insert_regs (cse.c:1029-1032, early bail :1018-1020) puts two pseudos in one equivalence class only when
  their MODES match, so an all-int copy-back is collapsed and swept, while the narrowed one is a truncation: no
  equivalence, the wide temp stays live and reaches reload as the move the target has. The fourth instruction is
  strength_reduce minting a shift giv from a wide counter whose every use is a cast; a HImode pseudo cannot be that giv.
  delever --propagate: 127 of 127 sibling(s) banked, 0 refused. 26,202 -> the census below.

- R20 narrow_chains: the agent PROVED the joint form is necessary, and the generator reproduces it. Single declarations
  scored 45/72/51/24, each chain alone 43, both chains together 0 — every intermediate worse than the search's own best
  of 11, so a beam over R12's one-declaration width moves cannot reach the answer from either side. Seven runs and 4,811
  compiles stalled at 11; R20 offers six candidates and the right one is a single compile.
- known-true check: run on b3's pre-bank text, R20's joint signed candidate scores 0 (MATCH) and its single-chain
  candidates score 43 and 51 — the agent's own hand-measured numbers, reproduced by the tool.
- chains are built conservatively from the body's text (two locals linked when one is assigned from the other, through a
  cast or a +/- constant), and only whole components are offered, so the partial narrowings the measurement showed are
  always worse are never generated. Selftest: the chain is found whole, an unlinked local is not pulled in, and a body
  with no linked pair offers nothing.

+++ .run/P36/agents/ov_SC04_011__func_8016CBC0/body.c
@@ -8,9 +8,9 @@ short func_8016CBC0(void)
   Blk_8016CBC0 *blk;
   Rec_8016CBC0 *sub;
   Rec_8016CBC0 *pv;
-  unsigned short uVar9;
+  short uVar9;
   short w9;
-  short uc;
+  unsigned short uc;
   unsigned int a2;
   int col;
   short iVar3;
@@ -50,7 +50,7 @@ short func_8016CBC0(void)
       {
         do
         {
-          pv = &sub[(short) uVar9];
+          pv = &sub[uVar9];
           {
             unsigned int a = pv->h[0];
             unsigned int b;
@@ -88,7 +88,7 @@ short func_8016CBC0(void)
           {
             uVar9 = 0xF;
           }
-          pv = &sub[(short) uVar9];
+          pv = &sub[uVar9];
           col = col - 0xF;
           a2 = (unsigned) (col & 0xFF);
           a2 = a2 >> 3;
